Add anchor room settings to menu.
On joining a room, a request for room settings is sent out. First player joining the room will fix the room's settings, and any subsequent joins will copy over the fixed settings.

Just to have a setting to tweak, this commit also adds a simple slider to adjust how many rupees need to be paid in order to teleport to another player.